# contracts/openapi/layer4-agents.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Layer 4: Agentic Workflow Engine",
    "description": "LangGraph-powered workflow orchestration for Value Fabric with multi-agent support",
    "version": "0.2.0"
  },
  "paths": {
...
    "/v1/users/invite": {
      "post": {
        "tags": [
          "Users"
        ],
        "summary": "Api Invite User",
        "description": "Invite a user to the caller's tenant. Requires ``tenant_admin`` role.",
        "operationId": "api_invite_user_v1_users_invite_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserInviteRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserModel"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/users": {
      "get": {
        "tags": [
          "Users"
        ],
        "summary": "Api List Users",
        "description": "List all users in the caller's tenant. Requires ``tenant_admin`` role.",
        "operationId": "api_list_users_v1_users_get",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 500,
              "minimum": 1,
              "default": 100,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/UserModel"
                  },
                  "title": "Response Api List Users V1 Users Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/users/{user_id}": {
      "get": {
        "tags": [
          "Users"
        ],
        "summary": "Api Get User",
        "description": "Get a user by ID. Requires ``tenant_admin`` role.",
        "operationId": "api_get_user_v1_users__user_id__get",
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "User Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserModel"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "Users"
        ],
        "summary": "Api Update User",
        "description": "Update a user's role or status. Requires ``tenant_admin`` role.",
        "operationId": "api_update_user_v1_users__user_id__patch",
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "User Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserUpdateRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserModel"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Users"
        ],
        "summary": "Api Deactivate User",
        "description": "Deactivate a user. Requires ``tenant_admin`` role.",
        "operationId": "api_deactivate_user_v1_users__user_id__delete",
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "User Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/v1/feature-flags": {
      "get": {
        "tags": [
          "Feature Flags"
        ],
        "summary": "List Feature Flags",
        "description": "List feature flags for the caller's tenant.",
        "operationId": "list_feature_flags_v1_feature_flags_get",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 500,
              "minimum": 1,
              "default": 100,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/FeatureFlagResponse"
                  },
                  "title": "Response List Feature Flags V1 Feature Flags Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/feature-flags/{flag_key}": {
      "get": {
        "tags": [
          "Feature Flags"
        ],
        "summary": "Get Feature Flag",
        "description": "Get a single feature flag by key.",
        "operationId": "get_feature_flag_v1_feature_flags__flag_key__get",
        "parameters": [
          {
            "name": "flag_key",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Flag Key"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FeatureFlagResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Feature Flags"
        ],
        "summary": "Upsert Feature Flag",
        "description": "Create or update a feature flag.",
        "operationId": "upsert_feature_flag_v1_feature_flags__flag_key__put",
        "parameters": [
          {
            "name": "flag_key",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Flag Key"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FeatureFlagUpsertRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FeatureFlagResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Feature Flags"
        ],
        "summary": "Delete Feature Flag",
        "description": "Delete a feature flag.",
        "operationId": "delete_feature_flag_v1_feature_flags__flag_key__delete",
        "parameters": [
          {
            "name": "flag_key",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Flag Key"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/feature-flags/{flag_key}/evaluate": {
      "get": {
        "tags": [
          "Feature Flags"
        ],
        "summary": "Evaluate Feature Flag",
        "description": "Evaluate a feature flag for the current user context.",
        "operationId": "evaluate_feature_flag_v1_feature_flags__flag_key__evaluate_get",
        "parameters": [
          {
            "name": "flag_key",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Flag Key"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Evaluate Feature Flag V1 Feature Flags  Flag Key  Evaluate Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/v1/tenants/{tenant_id}/users": {
      "get": {
        "tags": [
          "layer4-agents"
        ],
        "summary": "GET /v1/tenants/{tenant_id}/users",
        "description": "Discovered from router source (value-fabric\\layer4-agents\\src\\tenants\\api\\routes\\admin.py).",
        "operationId": "discovered_get_v1_tenants_tenant_id_users",
        "responses": {
          "200": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "FeatureFlagResponse": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "tenant_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tenant Id"
          },
          "flag_key": {
            "type": "string",
            "title": "Flag Key"
          },
          "enabled": {
            "type": "boolean",
            "title": "Enabled"
          },
          "rollout_percentage": {
            "type": "integer",
            "maximum": 100.0,
            "minimum": 0.0,
            "title": "Rollout Percentage"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "metadata": {
            "additionalProperties": true,
            "type": "object",
            "title": "Metadata"
          },
          "created_at": {
            "type": "string",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "title": "Updated At"
          },
          "updated_by": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Updated By"
          }
        },
        "type": "object",
        "required": [
          "id",
          "tenant_id",
          "flag_key",
          "enabled",
          "rollout_percentage",
          "description",
          "metadata",
          "created_at",
          "updated_at",
          "updated_by"
        ],
        "title": "FeatureFlagResponse",
        "description": "Feature flag read model."
      },
      "FeatureFlagUpsertRequest": {
        "properties": {
          "enabled": {
            "type": "boolean",
            "title": "Enabled"
          },
          "rollout_percentage": {
            "type": "integer",
            "maximum": 100.0,
            "minimum": 0.0,
            "title": "Rollout Percentage"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "metadata": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Metadata"
          }
        },
        "type": "object",
        "required": [
          "enabled",
          "rollout_percentage"
        ],
        "title": "FeatureFlagUpsertRequest",
        "description": "Payload to create or update a feature flag."
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "Role": {
        "type": "string",
        "enum": [
          "super_admin",
          "tenant_admin",
          "content_admin",
          "analyst",
          "read_only",
          "system"
        ],
        "title": "Role",
        "description": "Business roles used across all layers.\n\nHierarchy (highest → lowest):\n    super_admin > tenant_admin > content_admin > analyst > read_only\n\nThe ``system`` role is reserved for internal service-to-service calls."
      },
...
      "UserInviteRequest": {
        "properties": {
          "email": {
            "type": "string",
            "title": "Email",
            "description": "Email to invite"
          },
          "display_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Display Name"
          },
          "role": {
            "$ref": "#/components/schemas/Role",
            "default": "analyst"
          }
        },
        "type": "object",
        "required": [
          "email"
        ],
        "title": "UserInviteRequest",
        "description": "Invite a new user to the tenant (tenant_admin only)."
      },
      "UserModel": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "tenant_id": {
            "type": "string",
            "format": "uuid",
            "title": "Tenant Id",
            "description": "Owning tenant"
          },
          "email": {
            "type": "string",
            "title": "Email",
            "description": "Email address (unique within tenant)"
          },
          "display_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Display Name"
          },
          "role": {
            "$ref": "#/components/schemas/Role",
            "default": "analyst"
          },
          "status": {
            "$ref": "#/components/schemas/UserStatus",
            "default": "invited"
          },
          "last_login_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Login At"
          },
          "invited_by": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Invited By"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "tenant_id",
          "email"
        ],
        "title": "UserModel",
        "description": "Read-only representation of a user (returned by API — no password)."
      },
      "UserStatus": {
        "type": "string",
        "enum": [
          "invited",
          "active",
          "deactivated"
        ],
        "title": "UserStatus",
        "description": "Lifecycle status of a user."
      },
      "UserUpdateRequest": {
        "properties": {
          "display_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Display Name"
          },
          "role": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Role"
              },
              {
                "type": "null"
              }
            ]
          },
          "status": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/UserStatus"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "title": "UserUpdateRequest",
        "description": "Update a user's role or status."
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}